tests/wpt/tests/html/rendering/widgets/textarea-scrollbar-sizing-002.html

<!DOCTYPE html>
<link rel="help" href="https://html.spec.whatwg.org/C/#the-textarea-element-2">
<title>Test textarea width and height accounting for scrollbars, with vertical writing mode</title>
<meta name="author" title="Luke Warlow" href="mailto:lwarlow@igalia.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<style>textarea { writing-mode: vertical-lr }</style>
<textarea id="rows1" rows="1"></textarea>
<textarea id="rows1Hidden" rows="1" style="overflow-y: hidden"></textarea>
<textarea id="rows1Scroll" rows="1" style="overflow-y: scroll"></textarea>
<textarea id="cols3" cols="3"></textarea>
<textarea id="cols3Hidden" cols="3" style="overflow-x: hidden"></textarea>
<textarea id="cols3Scroll" cols="3" style="overflow-x: scroll"></textarea>
<div id="scrollableDiv" style="overflow:scroll"></div>

<script>
const scrollbarsHaveThickness = (scrollableDiv.offsetHeight != 0);
test(() => {
  assert_equals(rows1.offsetHeight, rows1Hidden.offsetHeight);
  assert_equals(rows1.offsetWidth, rows1Hidden.offsetWidth);
}, 'rows=1 doesnt include scrollbar thickness');

test(() => {
  assert_equals(rows1.offsetHeight, rows1Scroll.offsetHeight);
  if (scrollbarsHaveThickness) {
    assert_less_than(rows1.offsetWidth, rows1Scroll.offsetWidth);
  } else {
    assert_equals(rows1.offsetWidth, rows1Scroll.offsetWidth);
  }
}, 'rows=1 overflow scroll includes scrollbar thickness');

test(() => {
  assert_equals(cols3.offsetHeight, cols3Scroll.offsetHeight);
  if (scrollbarsHaveThickness) {
    assert_greater_than(cols3.offsetHeight, cols3Hidden.offsetHeight);
  } else {
    assert_equals(cols3.offsetHeight, cols3Hidden.offsetHeight);
  }
  assert_equals(cols3.offsetWidth, cols3Scroll.offsetWidth);
  assert_equals(cols3.offsetWidth, cols3Hidden.offsetWidth);
}, 'cols=3 includes scrollbar thickness');
</script>
