LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/textarea-scrollbar-sizing-002-expected.txt



FAIL rows=1 doesnt include scrollbar thickness assert_equals: expected 146 but got 161
PASS rows=1 overflow scroll includes scrollbar thickness
FAIL cols=3 includes scrollbar thickness assert_greater_than: expected a number greater than 42 but got 42
